refactor recipe base classes
