+++ bocadillo/recipes.py
@@ -1,4 +1,4 @@
-from .base import Applicable
+from bocadillo.base import Applicable
 from .hooks import HooksMixin
 from .routing import RoutingMixin
 from .templates import TemplatesMixin
