+++ bocadillo/routing/__init__.py
@@ -1,4 +1,4 @@
 from .exceptions import RouteDeclarationError
 from .route import Route
 from .router import Router
-from .mixins import RoutingMixin
+from .mixin import RoutingMixin
